all: simplify "(err error)" return values to error
The original use of named return parameters and unadorned return
statements followed a recommendation that was possibly misunderstood.
In later discussions there was the argument that these uses of
unadorned returns did not correspond to idiomatic Go.

This change adjusts "(err error)" return parameters to simple "error"s.

Also, in the Windows implementation of Close() the error is not
disregarded anymore.